chore: reorganise the logos filesystem
