+++ .gitignore
@@ -1,3 +1,3 @@
-logos/*/dist
+logos/dist
 node_modules/
 coverage
